test-25

+++ motorController.h
@@ -60,7 +60,7 @@ typedef enum {
 
 int speedSensorThreadFlag = 1;
 int fd;
-volatile int power = 0;
+volatile double power = 0.0;
 
 void Motor_Stop();
 void PCA9685_SetPWM(UBYTE channel, UWORD on, UWORD off);
